+++ Changes.md
@@ -7,7 +7,7 @@
 - Establish folder structure.
 - Concept testing.
 - Added Changes.md to track revisions.
-- Added README.md.
+- Added README.md to tell about the game.
 
 #### June 28, 2024
 
@@ -15,3 +15,10 @@
   - Added additional static dictionaries.
   - Added sound controls.
   - Code refactoring.
+
+#### June 28, 2024
+
+- V3
+  - Application name changed to Lexicon Quest.
+  - Implemented online API call for new terms.
+  - Code refactoring